Add the annotate in the density meth plot, but not for centromeres. TODO: centromeres should be find
gene_body_meth : gene methylation bins are averaged base on the sum of the bin sizes for each index(an array of sum of bin sizes for each bin is considered) ; For the bins with no C, None is added to the bin, this makes all the genes methylation status array to be the size of the bin num; The flanking regions which has an overlap with the next gene are excluded; For the negative strand genes all the methylation status arrays are reversed; assertions added for the methylation status sizes; tick params is active for the y axis, to check the methylation level of genes; genes_df is sorted and the reset_index is done.

input_parser: The new make_meth_string method is replaced, The meth_string is generated by using the pandas df and assigning the values in a bulk, this will speed up the method. The method is checked by comparing its results with previous method.